RenAIssance_VLM_OCR_Tushar_Tyagi/requirements.txt
transformers>=4.40
torch>=2.1
bitsandbytes>=0.43
accelerate>=0.30
qwen-vl-utils
jiwer>=3.0
Pillow>=10.0
pyspellchecker>=0.7.1
torchvision>=0.25.0
tiktoken>=0.12.0
sentencepiece>=0.2.1